requirements.txt
scikit-learn
numpy
jupyter
pandas
mlflow
cassandra-driver
PyYaml
tqdm
-e .
